terrain blockworld
This is the result of the first step for creating a blockworld level for beeatrice. I used the terrain to model a simple ground, planes to create a river, pro builder to create a fence, and simple cubes to create a dam. I also used homogeneous colours to avoid distractions.